Don't perform 64-bit acesses to the PLIC
Recent QEMU will fault for 8-byte accesses. Use a uint32_t instead of
size_t to avoid those problems.

+++ machine/fdt.c
@@ -439,10 +439,10 @@ static void plic_done(const struct fdt_scan_node *node, void *extra)
     if (hart < MAX_HARTS) {
       hls_t *hls = OTHER_HLS(hart);
       if (cpu_int == IRQ_M_EXT) {
-        hls->plic_m_ie     = ptr_to_ddccap((size_t*)((uintptr_t)scan->reg + ENABLE_BASE + ENABLE_SIZE * index));
+        hls->plic_m_ie     = ptr_to_ddccap((uint32_t*)((uintptr_t)scan->reg + ENABLE_BASE + ENABLE_SIZE * index));
         hls->plic_m_thresh = ptr_to_ddccap((uint32_t*) ((uintptr_t)scan->reg + HART_BASE   + HART_SIZE   * index));
       } else if (cpu_int == IRQ_S_EXT) {
-        hls->plic_s_ie     = ptr_to_ddccap((size_t*)((uintptr_t)scan->reg + ENABLE_BASE + ENABLE_SIZE * index));
+        hls->plic_s_ie     = ptr_to_ddccap((uint32_t*)((uintptr_t)scan->reg + ENABLE_BASE + ENABLE_SIZE * index));
         hls->plic_s_thresh = ptr_to_ddccap((uint32_t*) ((uintptr_t)scan->reg + HART_BASE   + HART_SIZE   * index));
       } else {
         printm("PLIC wired hart %d to wrong interrupt %d", hart, cpu_int);
